update for NumPy 2.0 (np.int -> np.int32, np.map -> np.array)

+++ elfpy/plots.py
@@ -128,7 +128,7 @@ def _get_data(dx, dy):
 
     length = dx.shape[0]
     num = int(np.ceil(float(length) / ic))
-    ii = np.linspace(0, length - 1, num, dtype=np.int)
+    ii = np.linspace(0, length - 1, num, dtype=np.int32)
 
     return dx[ii], dy[ii]
